% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncats.R
\name{cat_function}
\alias{cat_function}
\title{A Cat Function}
\usage{
cat_function(love = TRUE, return_cats = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}

\item{return_cats}{Return a cat. Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
